requirements.txt
async-lru
discord.py
jishaku
python-dotenv
supabase
toml
